Changed OnWork event handlers to use Int64 now

+++ Indy10/Lib/System/IdComponent.pas
@@ -97,11 +97,9 @@ type
     Level: Integer;
   end;
 
-  TWorkBeginEvent = procedure(ASender: TObject; AWorkMode: TWorkMode;
-   AWorkCountMax: Integer) of object;
+  TWorkBeginEvent = procedure(ASender: TObject; AWorkMode: TWorkMode; AWorkCountMax: Int64) of object;
   TWorkEndEvent = procedure(ASender: TObject; AWorkMode: TWorkMode) of object;
-  TWorkEvent = procedure(ASender: TObject; AWorkMode: TWorkMode;
-   AWorkCount: Integer) of object;
+  TWorkEvent = procedure(ASender: TObject; AWorkMode: TWorkMode; AWorkCount: Int64) of object;
 
   TIdComponent = class(TIdBaseComponent)
   protected
